Initial Commit
So far there is a rough template of the first screen. The buttons all
light up some-what correctly(highlighting needs work). The Portal,
Facebook, and Website operate as they should. All of the image assets
are in and there is an icon. This is mostly just an initial commit to
get the ball rolling.