More Structured Layout
The layout of the main screen is more structured. The buttons are
mostly where they need to be. There is text under them although the two
line text needs adjusting. Paging is implemented with page control. The
page control still needs to be able to tap the dots to page left and
right. The twitter feed is roughly in but no data yet. The bottom three
buttons, which go to websites, are working.
TODO:
*Dialog alternative for iOS
*Sections Call, Providers, Locations, Forms, Services, MSBHC Tracker,
FAQs, and Patient Resources
*Menu
*Twitter latest tweet
*Bus tracker stats
*Location Preference Dialog
*Page Control dots, paging by tapping
*Localization - Spanish
*Layouts specific to iPads and iPhones for better look.