Design Changes / iPhone Size Classes
Decided to pull the twitter bird from the feed area and into its own
button. Changes the main pages around. Now at the bottom of the first
page are faqs, patient resources, portal. On the second page is
Facebook, Twitter, Website.
Most of the iPhone Size classes are in. Just a few tweaks left.