Localization Configuration Almost Done
Checks put in sections where the text was too long for Spanish. In
these places, it shrinks the text to an appropriate size. The missing
translations were added back in but a few may still be missing. A bug
fixed with the providers section. A similar bug is in the form section
but will need to re-organize some code to fix properly. iPad size
classes are messed up and will need to be fixed in some areas. Code
still needs cleaned up in about half of the code.