Before Review
This is the app finished before the review. Documentation is finished.
The logo has been centered but slightly cuts off the options logo on
small screens. Bus tracker tested and working fine after a bug was
fixed. Splash screen changed to look a little better. Will need to add
in translations after review.