Review Changes & Misc
After the review with admin, there have been a few changes to the
design.
*The main screen has a white background title bar with the blue bar at
the bottom. This is to make it more uniform with the website.
*All the secondary screens have blue instead of teal backgrounds.
*Saturday changed to Saturday Clinic with vpchc green color.
*Providers Type “Medical” changed to “Primary Care” to match other
areas.
Added function that gets the current date in xx/xx/xxxx form and stores
it in a default when the bus update is ran. It then displays this on
the the bus page to give the user more input on having an outdated
schedule.
Some bug fixes with the provider page.